Add routing files in modules
