+++ src/app/modules/auth/auth.module.ts
@@ -1,24 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AuthComponent } from './auth.component';
-import { Routes, RouterModule } from '@angular/router';
-import { BookListComponent } from '@app/screens/book-list/book-list.component';
 import { AuthGuard } from '@app/guards/auth.guard';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: AuthComponent,
-    children: [
-      { path: 'books', component: BookListComponent }
-    ]
-  }
-];
-
+import { AuthRoutingModule } from './auth-routing.module';
 @NgModule({
   imports: [
     CommonModule,
-    RouterModule.forChild(routes)
+    AuthRoutingModule
   ],
   providers: [AuthGuard]
 })
